actions: Split out zerver.lib.sounds.
(cherry picked from commit 6a70f75587dcf5d12f1d7c0e9a9f39797dec8069)
