+++ zerver/lib/actions.py
@@ -2,7 +2,6 @@ import datetime
 import hashlib
 import itertools
 import logging
-import os
 import time
 from collections import defaultdict
 from dataclasses import asdict, dataclass, field
@@ -130,7 +129,6 @@ from zerver.lib.send_email import (
 )
 from zerver.lib.server_initialization import create_internal_realm, server_initialized
 from zerver.lib.sessions import delete_user_sessions
-from zerver.lib.storage import static_path
 from zerver.lib.stream_color import pick_colors
 from zerver.lib.stream_subscription import (
     SubInfo,
@@ -4323,22 +4321,6 @@ def send_peer_remove_events(
         altered_user_dict=altered_user_dict,
         private_peer_dict=private_peer_dict,
     )
-
-
-def get_available_notification_sounds() -> List[str]:
-    notification_sounds_path = static_path("audio/notification_sounds")
-    available_notification_sounds = []
-
-    for file_name in os.listdir(notification_sounds_path):
-        root, ext = os.path.splitext(file_name)
-        if "." in root:  # nocoverage
-            # Exclude e.g. zulip.abcd1234.ogg (generated by production hash-naming)
-            # to avoid spurious duplicates.
-            continue
-        if ext == ".ogg":
-            available_notification_sounds.append(root)
-
-    return sorted(available_notification_sounds)
 
 
 def notify_subscriptions_removed(
