+++ zerver/lib/events.py
@@ -12,7 +12,6 @@ from zerver.lib.actions import (
     default_stream_groups_to_dicts_sorted,
     do_get_streams,
     gather_subscriptions_helper,
-    get_available_notification_sounds,
     get_default_streams_for_realm,
     get_owned_bot_dicts,
     get_web_public_streams,
@@ -43,6 +42,7 @@ from zerver.lib.push_notifications import push_notifications_enabled
 from zerver.lib.realm_icon import realm_icon_url
 from zerver.lib.realm_logo import get_realm_logo_source, get_realm_logo_url
 from zerver.lib.soft_deactivation import reactivate_user_if_soft_deactivated
+from zerver.lib.sounds import get_available_notification_sounds
 from zerver.lib.stream_subscription import handle_stream_notifications_compatibility
 from zerver.lib.subscription_info import get_web_public_subs
 from zerver.lib.timestamp import datetime_to_timestamp
