+++ zerver/views/user_settings.py
@@ -26,7 +26,6 @@ from zerver.lib.actions import (
     do_change_user_setting,
     do_regenerate_api_key,
     do_start_email_change_process,
-    get_available_notification_sounds,
 )
 from zerver.lib.avatar import avatar_url
 from zerver.lib.email_validation import (
@@ -40,6 +39,7 @@ from zerver.lib.rate_limiter import RateLimitedUser
 from zerver.lib.request import REQ, has_request_variables
 from zerver.lib.response import json_success
 from zerver.lib.send_email import FromAddress, send_email
+from zerver.lib.sounds import get_available_notification_sounds
 from zerver.lib.upload import upload_avatar_image
 from zerver.lib.validator import check_bool, check_int, check_int_in, check_string_in
 from zerver.models import UserProfile, avatar_changes_disabled, name_changes_disabled
